requirements.txt
Django==5.2.5
djangorestframework==3.16.0
python-decouple==3.8
django-cors-headers==4.6.0